chore(deps): update TagLib to 2.1.1 (#4281)
* chore: update CROSS_TAGLIB_VERSION to 2.1.1-1

* feat: add run-docker target

Introduced a new Makefile target `run-docker` that allows users to run a Navidrome Docker image with specified tags. This addition simplifies the process of launching the Docker container by handling volume mappings for configuration and music folders. The change enhances the development workflow by making it easier to test and run PR images

Signed-off-by: Deluan <deluan@navidrome.org>

---------

Signed-off-by: Deluan <deluan@navidrome.org>